Refactor onboarding page and remove unused code
